requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
newspaper3k>=0.2.8
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.0
python-dateutil>=2.8.0
schedule>=1.2.0
click>=8.1.0